Moved nav data to navigation/ - update all imports after architecture revamp
